refactor(DateInput): migrate to TypeScript
Converts DateInput to TSX with DateInputProps extending TextInputProps
(minus the props DateInput redefines), drops PropTypes, and exports
DateInputProps from src/index.ts.

Notes on typing decisions (runtime unchanged):
- label is typed optional, unlike the old propTypes .isRequired:
  TableDateInput omits it and supplies a label via a custom
  renderInput, so required-label was already untrue in practice
- disableDates reuses flatpickr's Options["disable"] type
- isNaN(selectedDate) became isNaN(selectedDate.getTime()),
  which is the same coercion isNaN performed implicitly
